add review ok

+++ Models/CRM/CustomerReview.cs
@@ -11,7 +11,7 @@ namespace ERP_System_Project.Models.CRM
         public string Comment { get; set; }
 
         [Range(1, 5)]
-        public decimal Rating { get; set; } // e.g., 1 to 5
+        public int Rating { get; set; } // e.g., 1 to 5
         public  int CustomerId{ get; set; }
         public  Customer? Customer{ get; set; }
         public  int ProductId{ get; set; }
