+++ ViewModels/ECommerce/ProductDetailsVM.cs
@@ -1,4 +1,5 @@
 ﻿using ERP_System_Project.ViewModels.Inventory;
+using ERP_System_Project.ViewModels.CRM;
 
 namespace ERP_System_Project.ViewModels.ECommerce
 {
@@ -20,7 +21,7 @@ namespace ERP_System_Project.ViewModels.ECommerce
 
         public int QuantityInStock { get; set; }
 
-        public int TotalRate { get; set; } // 0 - 5 (stars)
+        public  double TotalRate { get; set; } // 0 - 5 (stars)
         public int NumberOfReviews { get; set; }
         public List<CustomerReviewVM> Reviews { get; set; } = new List<CustomerReviewVM>();
         public List<AttributeVM> Attributes { get; set; } = new List<AttributeVM>();
